auth/error: add to app router and extract to /modules
